[TASK] Add USD and EUR currency pages to Fund tab (RT-2876)
