+++ src/jade/tabs/banner/unfunded.jade
@@ -5,5 +5,7 @@ div.auth-attention.banner
       a(href="https://support.ripplelabs.com/hc/en-us/articles/202964876-Activating-Your-Wallet", target="_blank") Suggestions for getting XRP.
     li(l10n)
       a(href="#/btc") Deposit BTC
-    //li(l10n)
-      a(href="#/usd") Deposit USD (US users only)
+    li(l10n)
+      a(href="#/usd") make a deposit for at least 10 USD at SnapSwap.US
+    li(l10n)
+      a(href="#/eur") make a deposit for at least 10 EUR at SnapSwap.EU
