+++ src/jade/tabs/eur.jade
@@ -1,4 +1,4 @@
-section.col-xs-12.content(ng-controller='UsdCtrl')
+section.col-xs-12.content(ng-controller='EurCtrl')
 
   //- Disconnected
 
@@ -28,7 +28,7 @@ section.col-xs-12.content(ng-controller='UsdCtrl')
             h2.tx-status(l10n) There was an error while saving your changes.&#32;
             p(l10n) Please try again later.
     .col-sm-12
-      .alert.alert-warning(ng-show='account.Balance && !can_add_trust && !usdConnected', l10n) You must have at least
+      .alert.alert-warning(ng-show='account.Balance && !can_add_trust && !eurConnected', l10n) You must have at least
         strong  {{account.reserve_to_add_trust | rpamount:0}} XRP
         |  to add a new trust line.&#32;
         a(href="https://ripple.com/wiki/Reserves", target="_blank") More information
@@ -36,74 +36,69 @@ section.col-xs-12.content(ng-controller='UsdCtrl')
     .col-sm-3
       include fund/menu
     .col-sm-9.list(ng-hide='debug')
-      .nameLine(l10n) US Dollar Deposit
+      .nameLine(l10n) Euro Deposit
       .gateway.row
         .col-md-3
             .title(l10n) Gateway
-            .description(l10n) 
-              | SnapSwap US (~SnapSwap)
+            .description(l10n)
+              | SnapSwap EU (~SnapSwap)
               br
-              a(href="https://snapswap.us", target="_blank") www.snapswap.us
+              a(href="https://snapswap.eu", target="_blank") www.snapswap.eu
         .col-md-3
             .title(l10n) Availability
-            .description(l10n) USA
+            .description(l10n) All countries, except USA
         .col-md-3
             .title(l10n) Fees
-            .description(l10n) $ 0.31-$ 5.00&#32;
+            .description(l10n) 
+              p(l10n) € 0.31 – € 15.00&#32;
+              p(l10n)
+                a(href='https://snapswap.eu/#/fees', target="_blank") More details 
         .col-md-3
             .title(l10n) Rating
             .description(l10n)
-              a(href="https://coinist.co/ripple/assets/usd-snapswap-us", target="_blank") read reviews
+              a(href="https://coinist.co/ripple/assets/eur-snapswap-eu", target="_blank") read reviews
 
       .row.fund-tab-description
         .col-xs-12.col-md-8.col-lg-6
-          .descriptor(ng-show="!usdConnected", l10n) Ripple Trade has partnered with SnapSwap to provide easier access to USD. Following this action will enable SnapSwap to hold USD on your behalf.
+          .descriptor(ng-show="!eurConnected", l10n) Ripple Trade has partnered with SnapSwap to provide easier access to EUR. Following this action will enable SnapSwap to hold EUR on your behalf.
 
       .row.fund-tab-description
         .col-xs-7.col-sm-5.col-md-4
-          button.btn.btn-large(ng-show="!usdConnected && !loading", type="submit", ng-click="save_account()", ng-disabled="!account.Balance") Add SnapSwap.us
-          button.btn.btn-large(ng-show="!usdConnected && loading", type="submit", ng-disabled="loading") Adding...
-          button.btn.btn-large(ng-show="showInstructions && usdConnected", type="submit", ng-click="toggle_instructions()", l10n) Hide instructions
-          button.btn.btn-large(ng-show="usdConnected && !showInstructions", type="submit", ng-click="toggle_instructions()", l10n) Show instructions
+          button.btn.btn-large(ng-show="!eurConnected && !loading", type="submit", ng-click="save_account()", ng-disabled="!account.Balance") Add SnapSwap.eu
+          button.btn.btn-large(ng-show="!eurConnected && loading", type="submit", ng-disabled="loading") Adding...
+          button.btn.btn-large(ng-show="showInstructions && eurConnected", type="submit", ng-click="toggle_instructions()", l10n) Hide instructions
+          button.btn.btn-large(ng-show="eurConnected && !showInstructions", type="submit", ng-click="toggle_instructions()", l10n) Show instructions
 
         .connect.col-xs-5.col-sm-7.col-md-8(ng-hide="loading")
-          .description.fa.fa-times(ng-show="!usdConnected && account.Balance", l10n) Not connected
-          .description.fa.fa-check(ng-show="usdConnected && account.Balance", l10n) Connected
+          .description.fa.fa-times(ng-show="!eurConnected && account.Balance", l10n) Not connected
+          .description.fa.fa-check(ng-show="eurConnected && account.Balance", l10n) Connected
           .description(ng-show="!account.Balance", l10n) Your account has to be activated before you can add a gateway account.
 
-      .row(ng-show="showInstructions && usdConnected")
+      .row(ng-show="showInstructions && eurConnected")
         .instructions.col-md-10
           a.dismiss(href="", id="hide", ng-click="toggle_instructions()", l10n)  ×
-          .title(l10n) NEW CUSTOMERS- SIGN UP AT SNAPSWAP.US
+          .title(l10n) NEW CUSTOMERS- SIGN UP AT SNAPSWAP.EU
           .description(l10n) 
-            | Register and verify your account at SnapSwap US. You will need:&#32;
+            | Register and verify your account at SnapSwap EU. You will need:&#32;
             ul
               li
                 p(l10n) A valid government-issued ID: driver license, passport, ID card, or visa stamp.
               li
                 p(l10n) A proof of address: driver license, utility bill, bank statement, etc.
 
-          .title(l10n) EXISTING SNAPSWAP US CUSTOMERS
+          .title(l10n) EXISTING SNAPSWAP EU CUSTOMERS
           .description(l10n) 
             ul 
               li
                 p(l10n) Login to your account&#32;
-                  a(href="https://snapswap.us/#/signin", target="_blank") here.
+                  a(href="https://snapswap.eu/#/signin", target="_blank") here.
 
-          .title(l10n) DEPOSIT OPTIONS
+          .title(l10n) DEPOSITS AND WITHDRAWALS:
           .description(l10n)
             ul
               li
-                p(l10n) Deposit instantly via selected banks
+                p(l10n) Deposit from your bank (2-4 business days)
               li
-                p(l10n) Deposit cash at Bank of America branches (3 business days)
+                p(l10n) Withdraw to your bank account in SEPA (2-4 business days)
               li
-                p(l10n) Deposit via online bank transfer (3 business days)
-
-          .title(l10n) WITHDRAWAL OPTIONS
-          .description(l10n)
-            ul
-              li
-                p(l10n) Directly to your bank account (3-4 business days)
-              li
-                p(l10n) Directly to your PayPal account (several minutes)
+                p(l10n) Withdraw to SmartyCash Card (instant)
