+++ src/jade/tabs/fund/menu.jade
@@ -1,25 +1,27 @@
 .currencies.fund-menu
-  .first_menu_item(l10n) Ripple name
   a(href="#/xrp", ng-class="{active: $route.current.tabName == 'xrp'}", class="currency-xrp", l10n)
     i.icon.fa.fa-money.fa-xrp
     | XRP
-  .menu_item(l10n) Featured gateways
-  a(href="#/btc", ng-class="{active: $route.current.tabName == 'btc'}", class="currency-btc", l10n)
-    i.icon.fa.fa-money.fa-btc
-    | BTC
   a(href="#/brl", ng-class="{active: $route.current.tabName == 'brl'}", class="currency-brl", l10n)
     i.icon.fa.fa-money.fa-brl
     | BRL
-  //a(href="#/usd", ng-class="{active: $route.current.tabName == 'usd'}", l10n) USD (US only)
-  a(href="#/gold", ng-class="{active: $route.current.tabName == 'gold'}", class="currency-xau", l10n) 
-    i.icon.fa.fa-money.fa-xau
-    | XAU
+  a(href="#/btc", ng-class="{active: $route.current.tabName == 'btc'}", class="currency-btc", l10n)
+    i.icon.fa.fa-money.fa-btc
+    | BTC
+  a(href="#/eur", ng-class="{active: $route.current.tabName == 'eur'}", l10n)
+    i.icon.fa.fa-money.fa-eur
+    | EUR
   a(href="#/jpy", ng-class="{active: $route.current.tabName == 'jpy'}", class="currency-jpy", l10n)
     i.icon.fa.fa-money.fa-jpy
     | JPY
   a(href="#/mxn", ng-class="{active: $route.current.tabName == 'mxn'}", class="currency-mxn", l10n)
     i.icon.fa.fa-money.fa-mxn
     | MXN
-  .menu_item(l10n) Add any gateway
+  a(href="#/usd", ng-class="{active: $route.current.tabName == 'usd'}", l10n)
+    i.icon.fa.fa-money.fa-usd
+    | USD
+  a(href="#/gold", ng-class="{active: $route.current.tabName == 'gold'}", class="currency-xau", l10n) 
+    i.icon.fa.fa-money.fa-xau
+    | XAU
   a(href="#/trust", ng-class="{active: $route.current.tabName == 'trust'}", l10n) Gateways
 a(href="https://support.ripplelabs.com/hc/en-us/articles/202847686-Gateway-Information", target="_blank", l10n) Learn more about gateways
