+++ src/js/entry/web.js
@@ -91,7 +91,8 @@ var tabdefs = [
   require('../tabs/debug'),
   require('../tabs/btc'),
   require('../tabs/withdraw'),
-  // require('../tabs/usd'),
+  require('../tabs/usd'),
+  require('../tabs/eur'),
   require('../tabs/gold'),
   require('../tabs/tou'),
   require('../tabs/privacypolicy'),
