+++ src/less/ripple/tabs.less
@@ -2245,7 +2245,7 @@ header nav.navbar .open > .dropdown-menu {
   }
 }
 
-#t-fund, #t-xrp, #t-btc, #t-usd, #t-gold, #t-trust, #t-jpy, #t-brl, #t-mxn {
+#t-fund, #t-xrp, #t-btc, #t-usd, #t-gold, #t-trust, #t-jpy, #t-brl, #t-mxn, #t-eur {
 
   .currencies a {
     display: block;
@@ -2279,7 +2279,7 @@ header nav.navbar .open > .dropdown-menu {
   }
 }
 
-#t-fund, #t-xrp, #t-btc, #t-usd, #t-gold, #t-jpy, #t-brl, #t-mxn {
+#t-fund, #t-xrp, #t-btc, #t-usd, #t-gold, #t-jpy, #t-brl, #t-mxn, #t-eur {
     .B2R-wrap {
       font-size: 12px;
       .dismiss {
@@ -2304,7 +2304,6 @@ header nav.navbar .open > .dropdown-menu {
       }
   }
   .fund-tab-description {
-    margin-bottom: 20px;
     &.gateway {
       padding-left: 15px;
       padding-right: 15px;
@@ -2319,6 +2318,9 @@ header nav.navbar .open > .dropdown-menu {
     .description {
       font-size: 12px;
     }
+    .descriptor{
+      padding-bottom: 20px;  
+    }
     .connect {
       line-height: 42px;
       font-size: 12px;
@@ -2331,7 +2333,7 @@ header nav.navbar .open > .dropdown-menu {
     }
   }
 }
-#t-fund, #t-xrp, #t-btc, #t-usd, #t-gold, #t-brl, #t-jpy, #t-mxn  {
+#t-fund, #t-xrp, #t-btc, #t-usd, #t-gold, #t-brl, #t-jpy, #t-mxn, #t-eur {
   .currency {
     font-size: 18px;
     margin: 0 0 0 5px;
@@ -2646,7 +2648,7 @@ header nav.navbar .open > .dropdown-menu {
   }
 }
 
-#t-gold, #t-jpy, #t-btc, #t-brl, #t-mxn {
+#t-gold, #t-jpy, #t-btc, #t-brl, #t-mxn, #t-usd, #t-eur {
   .title {
     font-family: OpenSansBold;
     font-size: 12px;
@@ -2701,7 +2703,7 @@ header nav.navbar .open > .dropdown-menu {
     padding-left: 5px;
   }
 }
-#t-jpy {
+#t-jpy, #t-usd, #t-eur {
   .description {
     p {
       margin: 0;
